+++ projects/fusion-docs/src/app/pages/docs/docs.module.ts
@@ -8,10 +8,17 @@ import {environment} from '../../../environments/environment';
 import {StyleVersionButtonModule} from '../../components/style-version-button/style-version-button.module';
 import {MenuItemExampleModule} from '../../components/menu-item-example/menu-item-example.module';
 
+// for layout v4
+import {LayoutComponent} from '@ironsource/fusion-ui/components/layout/v4';
+import {TopFilterIncludeExcludeComponent} from '@ironsource/fusion-ui/components/top-filter-include-exclude';
+import {FormsModule, ReactiveFormsModule} from '@angular/forms';
+
 @NgModule({
     declarations: [DocsComponent],
     imports: [
         CommonModule,
+        FormsModule,
+        ReactiveFormsModule,
         DocsRoutingModule,
         LayoutV1Module.forRoot({svgOptions: {assetsPath: environment.assetsPath}}),
         // LayoutModule.forRoot({svgOptions: {assetsPath: environment.assetsPath}}),